Add logic to keep legacy logic
