+++ react-app/src/features/package/package-status/package-status.test.tsx
@@ -1,4 +1,5 @@
 import { UseQueryResult } from "@tanstack/react-query";
+import { screen } from "@testing-library/react";
 import { mockUseGetUser, TEST_PACKAGE_STATUS_ITEM } from "mocks";
 import { describe, expect, it, vi } from "vitest";
 
@@ -7,6 +8,13 @@ import * as api from "@/api/useGetUser";
 import { renderWithQueryClient } from "@/utils/test-helpers";
 
 import { PackageStatusCard } from ".";
+
+const mockUseFeatureFlag = vi.hoisted(() => vi.fn(() => false));
+
+vi.mock("@/hooks/useFeatureFlag", () => ({
+  useFeatureFlag: mockUseFeatureFlag,
+}));
+
 describe("Package Status test", () => {
   vi.spyOn(api, "useGetUser").mockImplementation(() => {
     const response = mockUseGetUser();
@@ -18,5 +26,20 @@ describe("Package Status test", () => {
       <PackageStatusCard submission={TEST_PACKAGE_STATUS_ITEM._source} />,
     );
     expect(asFragment()).toMatchSnapshot();
+  });
+
+  it("shows the SMART status without the OneMAC RAI withdraw sub-status when the flag is on", () => {
+    mockUseFeatureFlag.mockReturnValue(true);
+    const submission = {
+      ...TEST_PACKAGE_STATUS_ITEM._source,
+      cmsStatus: "Pending",
+      raiWithdrawEnabled: true,
+    };
+
+    renderWithQueryClient(<PackageStatusCard submission={submission} />);
+
+    expect(screen.getByText("Pending")).toBeInTheDocument();
+    expect(screen.queryByText("Withdraw Formal RAI Response - Enabled")).not.toBeInTheDocument();
+    mockUseFeatureFlag.mockReturnValue(false);
   });
 });
